% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.chron.R
\name{axis.chron}
\alias{axis.chron}
\title{Add axis with date/time labels for chron dimension}
\usage{
axis.chron(
  side,
  x,
  at,
  format,
  labels,
  tz = "GMT",
  lwd = graphics::par("lwd"),
  lwd.ticks = graphcis::par("lwd"),
  ...
)
}
\value{
location on tick marks
}
\description{
Adds axis in same way as call to 'axis'. Ticks and labels will be pretty date/time.
}
