% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.richardson.cbl.height.R
\name{bulk.richardson.cbl.height}
\alias{bulk.richardson.cbl.height}
\title{Calculate planetary boundary layer height}
\usage{
bulk.richardson.cbl.height(
  snd,
  ri.cr = 0.25,
  plot = TRUE,
  unit = c("ASL", "AGL"),
  zlim = c(0, 5000),
  smooth = FALSE
)
}
\arguments{
\item{snd}{data.frame with vertical sounding data. Needs to include the fields 'pt', 'sh',
'uu', 'vv', and 'zz' for potential temperature, specific humidity, 
west-east wind speed, south-north wind speed and height, respectively.}

\item{ri.cr}{Critical bulk Richardson number. Altitude with Richardson numbers above 
this value are considered outside planetary boundary layer.}

\item{plot}{(logical) if TRUE (default) the vertical profiles are plotted and the estimated
boundary layer height indicated}

\item{unit}{Unit of vertical coordinate. One of 'ASL' or 'AGL'.}

\item{zlim}{Vertical limits of profile plot.}

\item{smooth}{(logical) If TRUE, apply a 5-point running mean before calculating Richardson
number. Useful for noisy sounding data. Default is FALSE: no running mean.}
}
\value{
Planetary boundary layer height in units provided by 'unit'.
}
\description{
Calculates the height of the planetary boundary layer height using a bulk Richardson number
method.
}
