\name{constants}

\alias{constants}
\alias{R.star}
\alias{R.air}
\alias{R.h2o}
\alias{k.b}
\alias{N.A}
\alias{mu.air}
\alias{mu.h2o}
\alias{T.0}
\alias{PP.0}
\alias{F.0}
\alias{g.0}
\alias{c.p}
\alias{c.v}

\title{Physical constants}

\description{
Constants often needed for atmospheric calculations.
}

\usage{
R.star
R.air
R.h2o
k.b
N.A
mu.air
mu.h2o
T.0
PP.0
F.0
g.0
c.p
c.v
}

\details{
\tabular{lll}{
Symbol	\tab Description \tab Units \cr
R.star	\tab Universal gas constant \tab  J/K/mol \cr
R.air 	\tab Specific gas constant for dry air \tab J/kg/K \cr
R.h2o	\tab Specific gas constant for water vapor \tab J/kg/K \cr
k.b		\tab Boltzmann constant \tab J/K \cr
N.A		\tab Avogardo number (number of atoms/molecules in one mole) \tab 1/mol \cr
mu.air 	\tab Molecular weight of dry air \tab g/mol \cr
mu.h2o	\tab Molecular weidht of water \tab g/mol \cr
T.0		\tab Freezing point of water 	\tab K	\cr
PP.0	\tab Standard ambient sea level pressure \tab hPa \cr
F.0		\tab Solar constant 	\tab W/m2 \cr
g.0		\tab Average gravity of Earth 	\tab m/s2 \cr
c.p     \tab Specific heat capacity at constant pressure \tab J kg-1 K-1 \cr
c.v     \tab Specific heat capacity at constant volume   \tab J kg-1 K-1 \cr
}
}

\author{Stephan Henne \email{stephan.henne@empa.ch}}

\keyword{data}% at least one, from doc/KEYWORDS
