% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry.deposition.R
\name{dry.deposition.velocity}
\alias{dry.deposition.velocity}
\title{Calculate dry depostion volocity}
\usage{
dry.deposition.velocity(
  u.star,
  zz,
  z.0,
  GR,
  TT.s,
  lu = 2,
  season = 1,
  spec = c("O3", "SO2", "NO2", "NO", "HNO3", "H2O2", "HONO"),
  H.star,
  f.0,
  Dv.by.Dq
)
}
\arguments{
\item{u.star}{Friction velocity in the surface layer (m/s)}

\item{zz}{Reference height for which to calculate dry deposition velocity (m)}

\item{z.0}{Roughness length (m)}

\item{GR}{Global radiation (W/m2)}

\item{TT.s}{Surface temperature in (degree C)}

\item{lu}{Landuse/landcover class used by Wesely. 1) Urban land; 2) agricultural land; 3) range land;
4) deciduous forest; 5) coniferous forest; 6) mixed forest; 7) water, both salt and fresh; 
8) barren land, mostly desert; 9) nonforested wetland; 10) mixed agricultural and range land;
11) rocky open areas with low growing shrubs}

\item{season}{Integer 1 to 5 describing different seasons: 1) midsummer with lush vegetation; 
2) autum with unharvested cropland; 3) late autum afer frost, no snow; 
4) winter, snow on ground and subfreezing; 5) transitional spring with partially green short annuals}

\item{spec}{Species for which deposition velocity should be calculated. One of O3, SO2, NO2, NO, HNO3 H2O2, HONO.}

\item{H.star}{Effective Henry's law constant    (mol/l bar-1).     If not given standard value by species is given.}

\item{f.0}{Reactivity factor (unitless; range 0-1). If not given standard value by species is given.}

\item{Dv.by.Dq}{Ratio of the diffusion coefficient of water vapor to that of gas q. If not given standard value by species is given.}
}
\value{
dry deposition velocity in m/s
}
\description{
Calculates the dry depostion velocity of a gas following the parameterisation of Wesely. On 
input meteorological condtions, land cover type and season need to be provided.
}
\references{
Wesely, M. L., 1989: Parameterization of surface resistances to gaseous dry deposition in regional-scale numerical models. Atmospheric Environment (1967), 23, 1293-1304.
}
