% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gauss.plume.R
\name{gauss.plume}
\alias{gauss.plume}
\title{Gauss plume calculation}
\usage{
gauss.plume(x, y, z, Q, h.e, u, stability = "D", h.m = NULL)
}
\arguments{
\item{x}{Location in x direction for which to calculate the Gauss plume concentration, given in m from source.
Note that x,y,z can be arrays, but they should be of the same shape.}

\item{y}{Location in y direction for which to calculate the Gauss plume concentration, given in m from source.}

\item{z}{Location in z direction for which to calculate the Gauss plume concentration, given in m above ground.}

\item{Q}{emission flux of the source. Given in kg/s.}

\item{h.e}{effective emission height. Given in m above ground.}

\item{u}{wind speed in main wind direction (x) given in m/s.}

\item{stability}{Stability class following Pasquill definition.}

\item{h.m}{Boundary layer height given in m above ground. If missing or NA no boundary layer top is assumed.}
}
\value{
Concentration resulting from source strength Q, given in kg/m3
}
\description{
Calculates concentration of a Gaussian plume for a given location (x,y,z),
assuming that the source is located at x=0, y=0 and z=h.e and main wind only in x direction.
Stability classes following the Pasquill definition are used.
}

