% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rkpp.R
\name{get.MCM.photolysis.rates}
\alias{get.MCM.photolysis.rates}
\title{Calculate photolysis rates}
\usage{
get.MCM.photolysis.rates(dtm, lon = 0, lat = 0)
}
\arguments{
\item{dtm}{(chron) single date/time}

\item{lon}{(numeric)    longitude in degrees east}

\item{lat}{(numeric)    latitude in degrees north}
}
\value{
vector of photolysis rates using the MCM photolysis reaction definitions.
}
\description{
Applies the MCM/CRI parameterisation of photolysis rates based on location and time only. 
A total of 61 photolyiss rates are calculated. Refer to MCM/CRI for details.
}
