% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.radiation.R
\name{global.radiation}
\alias{global.radiation}
\title{Calculate clear sky global radiation at time and location}
\usage{
global.radiation(dtm = 0, lon = 0, lat = 0)
}
\arguments{
\item{dtm}{date/time as chron object}

\item{lon}{longitude in degree east}

\item{lat}{latitude in degree north}
}
\value{
global radidation in units W m-2
}
\description{
Using simplified equation based on soloar elevation angle as given in Stull 1982
}
