% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rkpp.R
\name{init.KPP}
\alias{init.KPP}
\title{Initialise KPP}
\usage{
init.KPP(init.var = NULL, var.default = 0, dt, tt = 270, init.gl = NULL)
}
\arguments{
\item{init.var}{(list) initial values of all species}

\item{var.default}{(numeric) inital value for species not in init.var}

\item{dt}{(numeric)    time step in seconds}

\item{tt}{(numeric)     temperature in K}

\item{init.gl}{(list) initial values of user defined globals}
}
\description{
Initial values for chemical compounds, time step, temperature and other global variables.
}
