\name{isa.zz.2.pp}
\alias{isa.zz.2.pp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Altitude to pressure for international standard atmosphere }
\description{
    Estimates the pressure for a given altitude above sea level and international standard atmospheric conditions.
}
\usage{
isa.zz.2.pp(zz = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{zz}{ (numeric) altitude above sea level in [m] }
}
\details{
}
\value{
    (numeric) pressure in [hPa]
}
\references{ Compare http://en.wikipedia.org/wiki/International\_Standard\_Atmosphere and Etling, D.: 1996. Theoretische Meteorologie. }
\author{ stephan.henne@empa.ch }
\seealso{ \code{\link{isa.pp.2.zz}}}
\keyword{ manip }

