% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reh.freq}
\alias{reh.freq}
\title{Frequency of dispersion category}
\format{
data.frame containing fields:
\describe{
  \item{freq}{frequency of dispersion category}
  \item{WS}{Wind speed, central value in units m s-1.}
  \item{WD}{Wind direction, central value.}
  \item{stability}{Pasquill stability category}
  \item{h.m}{mixing layer height; all values NA as not measured; can be set for supplying data.frame to 'average.gauss.plume.from.freq'}
}
}
\usage{
data(reh.freq)
}
\description{
Frequency of wind-speed, wind-direction, stability categories for the MeteoSwiss site
Reckenholz and the year 2015.
}
\keyword{datasets}
