% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reh.ts}
\alias{reh.ts}
\title{Time series of observed meteorology and stability category}
\format{
data.frame containing fields:
\describe{
  \item{dtm}{date/time chron object}
  \item{WS}{Wind speed in units m s-1.}
  \item{WD}{Wind direction, central value.}
  \item{stability}{Pasquill stability category}
  \item{TT}{2m ambient temperature in degree C}
  \item{RH}{2m relative humidity in \%}
}
}
\usage{
data(reh.ts)
}
\description{
Time series of meteorological observations and stability categories for the MeteoSwiss site
Reckenholz with hourly resolution and for the year 2015.
}
\keyword{datasets}
