% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rkpp.R
\name{set.KPP.globals}
\alias{set.KPP.globals}
\title{Set KPP global variables}
\usage{
set.KPP.globals(vals.gl, tt)
}
\arguments{
\item{vals.gl}{named list of global values passed to KPP}

\item{tt}{temperature; treated separately from other global variables}
}
\description{
Set KPP global variables
}
