% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.zenith.angle.R
\name{solar.azimuth.angle}
\alias{solar.azimuth.angle}
\title{Calculate solar azimth angle}
\usage{
solar.azimuth.angle(tm, lon = 0, lat = 0)
}
\arguments{
\item{tm}{date/time object (chron)}

\item{lon}{Longitude in degrees east}

\item{lat}{Latitude in degrees north}
}
\value{
Solar azimuth angle in radians
}
\description{
The solar azimuth angle is the azimuth (horizontal angle with respect to north) of the Sun's 
position. This horizontal coordinate defines the Sun's relative direction along the 
local horizon, whereas the solar zenith angle (or its complementary angle solar elevation) 
	defines the Sun's apparent altitude (see 'solar.zenith.angle).
}
