% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.zenith.angle.R
\name{solar.elevation.angle}
\alias{solar.elevation.angle}
\title{Calculate solar elevation angle}
\usage{
solar.elevation.angle(tm = 0, lon = 0, lat = 0)
}
\arguments{
\item{tm}{date/time object (chron)}

\item{lon}{Longitude in degrees east}

\item{lat}{Latitude in degrees north}
}
\value{
Solar elevation angle in radians
}
\description{
The solar elevation angle (or its complementary angle solar zenith 'solar.zenith.angle) 
	defines the Sun's apparent altitude. It is the angle between the sun's rays and local 
horizon. Maximum value at solar noon. Night-time values smaller 0.
}
