% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.zenith.angle.R
\name{solar.zenith.angle}
\alias{solar.zenith.angle}
\title{Calculate solar zenith angle}
\usage{
solar.zenith.angle(tm = 0, lon = 0, lat = 0)
}
\arguments{
\item{tm}{date/time object (chron)}

\item{lon}{Longitude in degrees east}

\item{lat}{Latitude in degrees north}
}
\value{
Solar zenith angle in radians
}
\description{
The solar zenith angle (or its complementary angle solar elevation 'solar.elevation.angle) 
	defines the Sun's apparent altitude. It is the angle between the sun's rays and the 
vertical direction. Minimum value at solar noon. Night-time values larger than pi/2.
}
