% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.plot.R
\name{ts.plot}
\alias{ts.plot}
\title{Simple time-series plot}
\usage{
ts.plot(
  X,
  para = names(X)[2],
  xlim,
  ylim = NULL,
  xlab,
  ylab = para,
  pch,
  lty,
  col,
  leg.pos = "topleft",
  stacked = 0,
  accumulated = FALSE,
  legend,
  dtm.col = "dtm",
  flag = NULL,
  bg.flag = NULL,
  col.par,
  color.palette = rainbow,
  scale = NULL,
  ...
)
}
\arguments{
\item{X}{data.frame containing time-series data}

\item{para}{parameter(s)/column(s) to plot}

\item{xlim}{2-element vector giving the limits of the x-axis. If missing determined from X.}

\item{ylim}{2-element vector giving the limits of the y-axis. Default NULL: determine
from data.}

\item{xlab}{Label for the x-axis. Missing by default: Using time range as axis label.}

\item{ylab}{Label for y-axis. Default is to use 'para'.}

\item{pch}{symbol index to be used for the plot; if missing set automatically}

\item{lty}{line type to be used for the plot; if missing set automatically}

\item{col}{color to be used for the plot; if missing set automatically}

\item{leg.pos}{Legend position. Passed to call to 'legend'. Default: 'topleft'}

\item{stacked}{Default (0) is to plot different parameters/columns in same plot with 
single y-axis. But giving a positive number to stacked, multiple parameters
will be plotted on top of each other in separate panels. The number determines
how many parameters are plotted per sub-plot.}

\item{accumulated}{If TRUE the sum of all parameters is formed and stacked surfaces are 
plotted instead of lines/symbols. Requires all variables to either be 
positive or negative.}

\item{legend}{Alternative legend text.}

\item{dtm.col}{Name of column in 'X' that contains the time variable. Default 'dtm'.}

\item{flag}{Name of column that contains flagging data. Differently flagged data points 
will be plotted with different symbols/colors.}

\item{bg.flag}{Value of special flag for background data.}

\item{col.par}{Additional secondary parameter that can be used for coloring time series.}

\item{color.palette}{Color palette function used for coloring time series.}

\item{scale}{Additional scaling factor for all time series data. Default is NULL: no 
scaling.}

\item{...}{Additional argument passed to other routines.}
}
\description{
Plots one or more variables of the passed data.frame against time.
}
