% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rkpp.R
\name{update.deposition}
\alias{update.deposition}
\title{Calculate tendency due to deposition}
\usage{
\method{update}{deposition}(spec, conc, vd = NULL, dt = 1, dz = 100)
}
\arguments{
\item{spec}{vector of species to be treated}

\item{conc}{vector of concentrations from which to calculated the tendencies}

\item{vd}{vector of dry deposition velocities (units m s-1)}

\item{dt}{time step in seconds}

\item{dz}{Vertical height of surface layer in  meters. Default is 100.}
}
\value{
Concentration tendency due to dry deposition
}
\description{
Calculate tendency due to deposition
}
