% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rkpp.R
\name{update.emissions}
\alias{update.emissions}
\title{Calculate tendency due to emissions}
\usage{
\method{update}{emissions}(
  spec,
  E = NULL,
  mu = NULL,
  dt = 1,
  V = 1,
  dtm = chron(0),
  hour.profile = NULL
)
}
\arguments{
\item{spec}{vector of species to be treated}

\item{E}{(list) emissions by name, units: kg/s}

\item{mu}{(list) molar mass by name, units: g/mole}

\item{dt}{(numeric) time step, units: s}

\item{V}{(numeric)    volume of box, units: m3}

\item{dtm}{date/time (chron)}

\item{hour.profile}{time of day profile of emission scaling factors}
}
\value{
Concentration tendency due to dry deposition
}
\description{
Calculate tendency due to emissions
}
