% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rkpp.R
\name{update.KPP.concentrations}
\alias{update.KPP.concentrations}
\title{Submit all concentrations to KPP}
\usage{
\method{update}{KPP.concentrations}(conc, dconc)
}
\arguments{
\item{conc}{concentrations passed to kpp}

\item{dconc}{concentration increment to be added to conc before passing to kpp}
}
\description{
Submit all concentrations to KPP
}
