% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UU_VV_2_WS_WD.R
\name{UU_VV_2_WS_WD}
\alias{UU_VV_2_WS_WD}
\title{Convert horizontal wind vector components to wind spped and direction}
\usage{
UU_VV_2_WS_WD(uu, vv)
}
\arguments{
\item{uu}{wind component in west-east direction}

\item{vv}{wind component in south-north direction}
}
\value{
list of WS (wind speed) and WD (wind direction). Units of WS will be the same as 
			'uu' and 'vv'.
}
\description{
Units for wind components 'uu' and 'vv' have to be the same. 
Units of returned wind speed will be accordingly. 
'uu' and 'vv' can be vectors or arrays, but have to be of the same shape.
}
