% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_WD_2_UU_VV.R
\name{WS_WD_2_UU_VV}
\alias{WS_WD_2_UU_VV}
\title{Convert horizontal wind speed and direction into vector components}
\usage{
WS_WD_2_UU_VV(ws, wd)
}
\arguments{
\item{ws}{wind speed}

\item{wd}{wind direction}
}
\value{
list of 'uu' (west-east wind component) and 'vv' (south-north wind component). 
		Units of WS will be the same as 'ws'.
}
\description{
Units of returned wind components will be the same as for passed wind speed 'ws'.
'ws' and 'wd' can be vectors or arrays, but have to be of the same shape.
}
