% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.chron.R
\name{axis.chron}
\alias{axis.chron}
\title{Add axis with date/time labels for chron dimension}
\usage{
axis.chron(
  side,
  x,
  at,
  format,
  labels,
  tz = "GMT",
  lwd = par("lwd"),
  lwd.ticks = par("lwd"),
  ...
)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be drawn on.  The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{x}{Alternative values of where axis ticks should be drawn.}

\item{at}{the points at which tick-marks are to be drawn.  Non-finite (infinite, ‘NaN’ or ‘NA’) values are omitted.  By default (when ‘NULL’) tickmark locations are computed, see ‘Details’ below.}

\item{format}{A format string for the conversion of chron to string. See '?strptime' for details.}

\item{labels}{this can either be a logical value specifying whether (numerical) annotations are to be made at the tickmarks, or a character or expression vector of labels to be placed at the tick points.  (Other objects are coerced by ‘as.graphicsAnnot’.)  If this is not logical, ‘at’ should also be supplied and of the same length.  If ‘labels’ is of length zero after coercion, it has the same effect as supplying ‘TRUE’.}

\item{tz}{time zone of chron object. Default is "GMT".}

\item{lwd}{line width of axis line}

\item{lwd.ticks}{line width of tick marks}

\item{...}{other parameters passed to 'axis'}
}
\value{
location on tick marks
}
\description{
Adds axis in same way as call to 'axis'. Ticks and labels will be pretty date/time.
}
