% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.skewt.R
\name{skewT.plot}
\alias{skewT.plot}
\title{Draw a Skew-T diagram}
\usage{
skewT.plot(
  snd,
  T.surf,
  qv.surf,
  plot = TRUE,
  plot.cape = TRUE,
  plot.LCL = TRUE,
  dp = 100,
  ...
)
}
\arguments{
\item{snd}{(data.frame) should contain at least the columns 'PRES' for atmospheric pressure in 
hPa, 'TEMP' for atmospheric temperature in degree C, and 'MIXR' for the water vapour 
mixing ratio in g/kg.}

\item{T.surf}{optional surface temperature (degree C) used for LCL and CAPE calculation}

\item{plot.cape}{(logical) if TRUE CAPE is calculated and displayed}

\item{plot.LCL}{(logical) if TRUE lifting condensation level is calculated and displayed}

\item{pv.surf}{optional water vapor mixing ratio at the surface (g/kg) used for LCL and CAPE
calculation}
}
\value{
NULL at success
}
\description{
Draws a Skew-T diagram for a given atmsopheric profile of atmospheric temperature, specific
humidity and pressure. Optionally calculates the Lifting Condensation Level and Convective 
Available Potential Energy (CAPE).
}
\references{
Rogers&Yau
}
\author{
stephan.henne@empa.ch
}
