\name{solar.zenith.angle}
\alias{solar.zenith.angle}
\title{ Calculate solor zenith angle }
\description{
    Calculates the sun zenith angle from time and lat/lng information.
}
\usage{
solar.zenith.angle(tm = 0, lng = 0, lat = 0)
}
\arguments{
  \item{tm}{ (chron) time }
  \item{lng}{ (numeric) longitude }
  \item{lat}{ (numeric) latitude }
}
\value{
    solor zenith angle in radians
}
\references{ following Madronich1999a }
\author{ stephan.henne@empa.ch }
\keyword{ manip }

